+++ src/Oro/Bundle/ActionBundle/Controller/AjaxController.php
@@ -2,125 +2,108 @@
 
 namespace Oro\Bundle\ActionBundle\Controller;
 
-use Symfony\Bundle\FrameworkBundle\Controller\Controller;
 use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\HttpFoundation\RedirectResponse;
+
+use Symfony\Bundle\FrameworkBundle\Controller\Controller;
 
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Route;
+use Sensio\Bundle\FrameworkExtraBundle\Configuration\Method;
 
-use Doctrine\Common\Collections\ArrayCollection;
 use Doctrine\Common\Collections\Collection;
 
-use Oro\Bundle\ActionBundle\Exception\ForbiddenOperationException;
-use Oro\Bundle\ActionBundle\Exception\OperationNotFoundException;
-use Oro\Bundle\ActionBundle\Helper\ContextHelper;
-use Oro\Bundle\ActionBundle\Model\ActionData;
-use Oro\Bundle\ActionBundle\Model\Operation;
-use Oro\Bundle\ActionBundle\Model\OperationRegistry;
-
 use Oro\Bundle\SecurityBundle\Annotation\AclAncestor;
+use Oro\Bundle\ActionBundle\Model\Operation;
+use Oro\Bundle\ActionBundle\Handler\ExecuteOperationResult;
 
 class AjaxController extends Controller
 {
     /**
      * @Route("/operation/execute/{operationName}", name="oro_action_operation_execute")
      * @AclAncestor("oro_action")
+     * @Method({"POST"})
      *
      * @param Request $request
-     * @param string $operationName
+     * @param string  $operationName
+     *
      * @return Response
      */
-    public function executeAction(Request $request, $operationName)
+    public function executeAction(Request $request, $operationName): Response
     {
-        $data = $this->getContextHelper()->getActionData();
+        $operation = $this->get('oro_action.operation_registry')->findByName($operationName);
+        if (!$operation instanceof Operation) {
+            $message = sprintf('Operation with name "%s" not found', $operationName);
 
-        $errors = new ArrayCollection();
-        $code = Response::HTTP_OK;
-        $message = '';
-        $pageReload = true;
-
-        try {
-            $operation = $this->getOperationRegistry()->findByName($operationName);
-
-            if (!$operation instanceof Operation) {
-                throw new OperationNotFoundException($operationName);
-            }
-            if (!$operation->isAvailable($data, $errors)) {
-                throw new ForbiddenOperationException();
+            $routeName = $request->get('route');
+            if (null !== $routeName && !$request->isXmlHttpRequest()) {
+                return $this->handleFailedNonAjaxResponse($message, $routeName);
             }
 
-            $operation->execute($data, $errors);
-            $pageReload = $operation->getDefinition()->isPageReload();
-        } catch (OperationNotFoundException $e) {
-            $code = Response::HTTP_NOT_FOUND;
-        } catch (ForbiddenOperationException $e) {
-            $code = Response::HTTP_FORBIDDEN;
-        } catch (\Exception $e) {
-            $code = Response::HTTP_INTERNAL_SERVER_ERROR;
+            return new JsonResponse(
+                [
+                    'message' => $message,
+                    'success' => false
+                ],
+                Response::HTTP_NOT_FOUND
+            );
         }
+        $executionResult = $this->get('oro_action.handler.execute_operation')->process($operation);
 
-        if (isset($e)) {
-            $message = $e->getMessage();
-        }
-
-        return $this->handleResponse($request, $data, $code, $message, $errors, $pageReload);
+        return $this->handleExecutionResult($executionResult, $request);
     }
 
     /**
-     * @param Request $request
-     * @param ActionData $data
-     * @param int $code
-     * @param string $message
-     * @param Collection $errorMessages
-     * @param bool $pageReload
+     * @param ExecuteOperationResult $result
+     * @param Request                $request
+     *
      * @return Response
+     * @throws \InvalidArgumentException
      */
-    protected function handleResponse(
-        Request $request,
-        ActionData $data,
-        $code,
-        $message,
-        Collection $errorMessages,
-        bool $pageReload
-    ) {
-        $response = [
-            'success' => $code === Response::HTTP_OK,
-            'message' => $message,
-            'messages' => $this->prepareMessages($errorMessages),
-            'pageReload' => $pageReload
+    protected function handleExecutionResult(ExecuteOperationResult $result, Request $request): Response
+    {
+        $validationErrors = $result->getValidationErrors();
+        $actionData = $result->getActionData();
+        $success = $result->isSuccess();
+        $response   = [
+            'success'    => $success,
+            'message'    => $result->getExceptionMessage(),
+            'messages'   => $this->prepareMessages($validationErrors),
+            'pageReload' => $result->isPageReload()
         ];
-
-        // handle redirect for failure response on non ajax requests
-        if (!$request->isXmlHttpRequest() && !$response['success'] && null !== ($routeName = $request->get('route'))) {
-            $this->get('session')->getFlashBag()->add('error', $response['message']);
-
-            return $this->redirect($this->generateUrl($routeName));
-        }
-
-        if ($data->getRefreshGrid() || !$response['success'] || !$response['pageReload']) {
-            $response['refreshGrid'] = $data->getRefreshGrid();
-            $response['flashMessages'] = $this->get('session')->getFlashBag()->all();
-        } elseif ($data->getRedirectUrl()) {
-            if ($request->isXmlHttpRequest()) {
-                $response['redirectUrl'] = $data->getRedirectUrl();
-            } else {
-                return $this->redirect($data->getRedirectUrl());
+        if (!$success) {
+            $response['refreshGrid'] = $actionData->getRefreshGrid();
+            $routeName = $request->get('route');
+            if (null !== $routeName && !$request->isXmlHttpRequest()) {
+                return $this->handleFailedNonAjaxResponse($response['message'], $routeName);
+            }
+        } else {
+            if (!$response['pageReload'] || $actionData->getRefreshGrid()) {
+                $response['refreshGrid'] = $actionData->getRefreshGrid();
+                $response['flashMessages'] = $this->get('session')->getFlashBag()->all();
+            } elseif ($actionData->getRedirectUrl()) {
+                if ($request->isXmlHttpRequest()) {
+                    $response['redirectUrl'] = $actionData->getRedirectUrl();
+                } else {
+                    return $this->redirect($actionData->getRedirectUrl());
+                }
             }
         }
 
-        return new JsonResponse($response, $code);
+        return new JsonResponse($response, $result->getCode());
     }
 
     /**
      * @param Collection $messages
+     *
      * @return array
+     * @throws \InvalidArgumentException
      */
-    protected function prepareMessages(Collection $messages)
+    protected function prepareMessages(Collection $messages): array
     {
         $translator = $this->get('translator');
         $result = [];
-
         foreach ($messages as $message) {
             $result[] = $translator->trans($message['message'], $message['parameters']);
         }
@@ -129,18 +112,17 @@ class AjaxController extends Controller
     }
 
     /**
-     * @return OperationRegistry
+     * Handle failed response non ajax requests
+     *
+     * @param string $message
+     * @param string $routeName
+     *
+     * @return RedirectResponse
      */
-    protected function getOperationRegistry()
+    protected function handleFailedNonAjaxResponse(string $message, string $routeName): RedirectResponse
     {
-        return $this->get('oro_action.operation_registry');
-    }
+        $this->get('session')->getFlashBag()->add('error', $message);
 
-    /**
-     * @return ContextHelper
-     */
-    protected function getContextHelper()
-    {
-        return $this->get('oro_action.helper.context');
+        return $this->redirect($this->generateUrl($routeName));
     }
 }