+++ src/Oro/Bundle/ActivityListBundle/EventListener/SegmentConditionBuilderOptionsListener.php
@@ -14,10 +14,8 @@ class SegmentConditionBuilderOptionsListener
         $event->setOptions(array_merge_recursive(
             $event->getOptions(),
             [
-                'onFieldsUpdate' => [
-                    'toggleCriteria' => [
-                        'condition-activity',
-                    ],
+                'fieldsRelatedCriteria' => [
+                    'condition-activity',
                 ],
             ]
         ));