+++ src/Oro/Bundle/DataGridBundle/Resources/public/js/datagrid/page-size.js
@@ -139,6 +139,8 @@ define([
                 this.$el.hide();
             }
 
+            this.initControls();
+
             return this;
         }
     });