+++ src/Oro/Component/MessageQueue/Job/JobProcessor.php
@@ -4,9 +4,23 @@ namespace Oro\Component\MessageQueue\Job;
 use Oro\Component\MessageQueue\Client\Message;
 use Oro\Component\MessageQueue\Client\MessagePriority;
 use Oro\Component\MessageQueue\Client\MessageProducerInterface;
+use Oro\Component\MessageQueue\Provider\JobConfigurationProviderInterface;
+use Oro\Component\MessageQueue\Provider\NullJobConfigurationProvider;
 
+/**
+ * JobProcessor is a main class responsible for processing jobs, shifting it's responsibilities to other classes
+ * is quite difficult and would make it less readable.
+ * @SuppressWarnings(PHPMD.TooManyPublicMethods)
+ * @SuppressWarnings(PHPMD.TooManyMethods)
+ * @SuppressWarnings(PHPMD.ExcessiveClassComplexity)
+ */
 class JobProcessor
 {
+    /**
+     * @var JobConfigurationProviderInterface
+     */
+    private $jobConfigurationProvider;
+
     /**
      * @var JobStorage
      */
@@ -25,6 +39,28 @@ class JobProcessor
     {
         $this->jobStorage = $jobStorage;
         $this->producer = $producer;
+    }
+
+    /**
+     * @param JobConfigurationProviderInterface $jobConfigurationProvider
+     * @return $this
+     */
+    public function setJobConfigurationProvider(JobConfigurationProviderInterface $jobConfigurationProvider)
+    {
+        $this->jobConfigurationProvider = $jobConfigurationProvider;
+
+        return $this;
+    }
+
+    /**
+     * @return JobConfigurationProviderInterface
+     */
+    public function getJobConfigurationProvider()
+    {
+        if (!$this->jobConfigurationProvider) {
+            return new NullJobConfigurationProvider();
+        }
+        return $this->jobConfigurationProvider;
     }
 
     /**
@@ -77,17 +113,47 @@ class JobProcessor
         $job->setStatus(Job::STATUS_NEW);
         $job->setName($jobName);
         $job->setCreatedAt(new \DateTime());
+        $job->setLastActiveAt(new \DateTime());
         $job->setStartedAt(new \DateTime());
         $job->setJobProgress(0);
         $job->setUnique((bool) $unique);
 
+        return $this->saveJobAndStaleDuplicateIfQualifies($job);
+    }
+
+    /**
+     * @param Job $job
+     * @return null|Job
+     */
+    private function saveJobAndStaleDuplicateIfQualifies($job)
+    {
         try {
             $this->jobStorage->saveJob($job);
+            return $job;
         } catch (DuplicateJobException $e) {
-            return null;
+            $currentRootJob = $this->findRootJobByJobNameAndStatuses($job->getName(), Job::$activeStatuses);
+            if ($currentRootJob && $this->isJobStale($currentRootJob)) {
+                $this->staleRootJobAndChildren($currentRootJob);
+                return $this->saveJobAndStaleDuplicateIfQualifies($job);
+            }
         }
+        return null;
+    }
 
-        return $job;
+    /**
+     * @param Job $job
+     * @return bool
+     */
+    private function isJobStale(Job $job)
+    {
+        if ($job->getStatus() === Job::STATUS_STALE) {
+            return true;
+        }
+        $timeBeforeStale = $this->getJobConfigurationProvider()->getTimeBeforeStaleForJobName($job->getName());
+        if ($timeBeforeStale !== null && $timeBeforeStale != -1) {
+            return $job->getLastActiveAt() <= new \DateTime('- ' . $timeBeforeStale. ' seconds');
+        }
+        return false;
     }
 
     /**
@@ -150,6 +216,7 @@ class JobProcessor
         $job->setStartedAt(new \DateTime());
 
         $this->jobStorage->saveJob($job);
+        $this->updateJobLastActiveAtAndSave($job->getRootJob());
 
         $this->producer->send(Topics::CALCULATE_ROOT_JOB_STATUS, [
             'jobId' => $job->getId()
@@ -179,12 +246,44 @@ class JobProcessor
         $job->setJobProgress(1);
         $job->setStoppedAt(new \DateTime());
         $this->jobStorage->saveJob($job);
-
+        $this->updateJobLastActiveAtAndSave($job->getRootJob());
         $this->producer->send(Topics::CALCULATE_ROOT_JOB_STATUS, [
             'jobId' => $job->getId()
         ]);
 
         $this->sendRecalculateJobProgressMessage($job);
+    }
+
+    /**
+     * @param Job $job
+     */
+    private function updateJobLastActiveAtAndSave(Job $job)
+    {
+        $job->setLastActiveAt(new \DateTime());
+        $this->jobStorage->saveJob($job);
+    }
+    
+    /**
+     * @param Job $rootJob
+     */
+    private function staleRootJobAndChildren(Job $rootJob)
+    {
+        if (!$rootJob->isRoot()) {
+            throw new \LogicException(sprintf('Can\'t stale child jobs. id: "%s"', $rootJob->getId()));
+        }
+
+        $this->jobStorage->saveJob($rootJob, function (Job $rootJob) {
+            $rootJob->setStatus(Job::STATUS_STALE);
+            $rootJob->setStoppedAt(new \DateTime());
+
+            foreach ($rootJob->getChildJobs() as $childJob) {
+                if (in_array($childJob->getStatus(), Job::$activeStatuses)) {
+                    $childJob->setStatus(Job::STATUS_STALE);
+                    $childJob->setStoppedAt(new \DateTime());
+                    $this->jobStorage->saveJob($childJob);
+                }
+            }
+        });
     }
 
     /**
@@ -210,6 +309,7 @@ class JobProcessor
         $job->setStoppedAt(new \DateTime());
 
         $this->jobStorage->saveJob($job);
+        $this->updateJobLastActiveAtAndSave($job->getRootJob());
 
         $this->producer->send(Topics::CALCULATE_ROOT_JOB_STATUS, [
             'jobId' => $job->getId()
@@ -239,6 +339,7 @@ class JobProcessor
 
         $job->setStatus(Job::STATUS_FAILED_REDELIVERED);
         $this->jobStorage->saveJob($job);
+        $this->updateJobLastActiveAtAndSave($job->getRootJob());
 
         $this->producer->send(Topics::CALCULATE_ROOT_JOB_STATUS, [
                 'jobId' => $job->getId()
@@ -256,7 +357,7 @@ class JobProcessor
 
         $job = $this->jobStorage->findJobById($job->getId());
 
-        if (! in_array($job->getStatus(), [Job::STATUS_NEW, Job::STATUS_RUNNING, Job::STATUS_FAILED_REDELIVERED])) {
+        if (! in_array($job->getStatus(), Job::$activeStatuses)) {
             throw new \LogicException(sprintf(
                 'Can cancel only new or running jobs. id: "%s", status: "%s"',
                 $job->getId(),
@@ -272,6 +373,7 @@ class JobProcessor
         }
 
         $this->jobStorage->saveJob($job);
+        $this->updateJobLastActiveAtAndSave($job->getRootJob());
 
         $this->producer->send(Topics::CALCULATE_ROOT_JOB_STATUS, [
             'jobId' => $job->getId()
@@ -290,10 +392,7 @@ class JobProcessor
         }
 
         foreach ($job->getChildJobs() as $childJob) {
-            if (in_array(
-                $childJob->getStatus(),
-                [Job::STATUS_NEW, Job::STATUS_RUNNING, Job::STATUS_FAILED_REDELIVERED]
-            )) {
+            if (in_array($childJob->getStatus(), Job::$activeStatuses)) {
                 $this->cancelChildJob($childJob);
             }
         }